V.1.3.0 - Usability update

+++ src/com/keitaproject/Lesson3/controller/PersonController.java
@@ -1,14 +1,16 @@
 package com.keitaproject.Lesson3.controller;
+import com.keitaproject.Lesson3.Exceptions.IfListEmpty;
+import com.keitaproject.Lesson3.Exceptions.IllegalVolumeDelete;
 import com.keitaproject.Lesson3.model.Person;
 import com.keitaproject.Lesson3.service.PersonService;
 import com.keitaproject.Lesson3.service.PersonServiceImpl;
 import com.keitaproject.Lesson3.storage.PersonStorageImpl;
-
 import java.time.LocalDate;
 import java.time.Period;
 import java.time.format.DateTimeParseException;
+import java.util.Comparator;
+import java.util.InputMismatchException;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Scanner;
 public class PersonController {
     private final Scanner scanner = new Scanner(System.in);
@@ -37,17 +39,33 @@ public class PersonController {
             }
         }
     }
-    private void add() {
+    public void add() {
         try {
             scanner.nextLine();
             System.out.println("Введите имя:");
             String firstName = scanner.nextLine();
+            while (firstName.matches(".*\\d.*") || firstName.contains(" ")) {
+            System.out.println("Имя не должно содержать цифры и пробелы. Пожалуйста, введите имя ещё раз:");
+            firstName = scanner.nextLine();
+        }
             System.out.println("Введите фамилию:");
             String lastName = scanner.nextLine();
+            while (lastName.matches(".*\\d.*") || lastName.contains(" ")) {
+            System.out.println("Фамилия не должна содержать цифры и пробелы. Пожалуйста, введите фамилию ещё раз:");
+            lastName = scanner.nextLine();
+        }
             System.out.println("Введите отчество:");
             String middleName = scanner.nextLine();
+            while (middleName.matches(".*\\d.*") || middleName.contains(" ")) {
+            System.out.println("Отчество не должно содержать цифры и пробелы. Пожалуйста, введите отчество ещё раз:");
+            middleName = scanner.nextLine();
+        }
             System.out.println("Введите дату рождения в формате гггг-мм-дд:");
             String dateOfBirth = scanner.nextLine();
+            while (!dateOfBirth.matches("^[0-9]{4}-[0-9]{2}-[0-9]{2}$") || dateOfBirth.contains(" ")) {
+            System.out.println("Дата рождения должна содержать только цифры в формате гггг-мм-дд и не должна содержать пробелы. Пожалуйста, введите дату рождения ещё раз:");
+            dateOfBirth = scanner.nextLine();
+        }
             LocalDate.parse(dateOfBirth);
             int age = Period.between(LocalDate.parse(dateOfBirth), LocalDate.now()).getYears();
             if (age < 0) {
@@ -55,30 +73,54 @@ public class PersonController {
             }
             personService.add(firstName, lastName, middleName, age, dateOfBirth);
             System.out.println("Запись добавлена.");
-        } catch (DateTimeParseException e) {
+    } catch (DateTimeParseException e) {
             System.out.println("Произошла ошибка: Неправильный формат даты рождения. Запись не добавлена.");
-        } catch (Exception e) {
+    } catch (Exception e) {
             System.out.println("Произошла ошибка: " + e.getMessage() + ". Запись не добавлена.");
         }
     }
     private void update() {
         try {
+            List<Person> personList = personService.list();
+            if (personList.isEmpty()) {
+                throw new IfListEmpty("Список пользователей пуст!");
+            }
             List<Person> persons = personService.findAll();
-            int maxId = persons.size() - 1;
-            System.out.println("Введите id записи для редактирования (от 0 до " + maxId + "):");
+            int maxId = persons.size();
+            System.out.print("Доступные id пользователей для редактирования (от 1 до " + maxId + ")" +  ".\nID после двоеточия вы можете использовать для редактирования: ");
+            for (Person person : persons) {
+                System.out.print(person.getId() + " ");
+            }
+            System.out.println();
             int id = scanner.nextInt();
             scanner.nextLine();
             if (id < 0 || id > maxId) {
                 throw new IllegalArgumentException("Введенный id не существует");
             }
-            System.out.println("Введите новое имя:");
+            System.out.println("Введите имя:");
             String firstName = scanner.nextLine();
-            System.out.println("Введите новую фамилию:");
+            while (firstName.matches(".*\\d.*") || firstName.contains(" ")) {
+            System.out.println("Имя не должно содержать цифры и пробелы. Пожалуйста, введите имя ещё раз:");
+            firstName = scanner.nextLine();
+            }
+            System.out.println("Введите фамилию:");
             String lastName = scanner.nextLine();
-            System.out.println("Введите новое отчество:");
+            while (lastName.matches(".*\\d.*") || lastName.contains(" ")) {
+            System.out.println("Фамилия не должна содержать цифры и пробелы. Пожалуйста, введите фамилию ещё раз:");
+            lastName = scanner.nextLine();
+            }
+            System.out.println("Введите отчество:");
             String middleName = scanner.nextLine();
+            while (middleName.matches(".*\\d.*") || middleName.contains(" ")) {
+            System.out.println("Отчество не должно содержать цифры и пробелы. Пожалуйста, введите отчество ещё раз:");
+            middleName = scanner.nextLine();
+            }
             System.out.println("Введите дату рождения в формате гггг-мм-дд:");
             String dateOfBirth = scanner.nextLine();
+            while (!dateOfBirth.matches("^[0-9]{4}-[0-9]{2}-[0-9]{2}$") || dateOfBirth.contains(" ")) {
+            System.out.println("Дата рождения должна содержать только цифры в формате гггг-мм-дд и не должна содержать пробелы. Пожалуйста, введите дату рождения ещё раз:");
+            dateOfBirth = scanner.nextLine();
+            }
             LocalDate.parse(dateOfBirth);
             int age = Period.between(LocalDate.parse(dateOfBirth), LocalDate.now()).getYears();
             if (age < 0) {
@@ -87,10 +129,11 @@ public class PersonController {
             personService.update(id, firstName, lastName, middleName, age, dateOfBirth);
             System.out.println("Запись изменена.");
         } catch (DateTimeParseException e) {
-            System.out.println("Произошла ошибка: Неправильный формат даты рождения. Запись не добавлена.");
-        } catch (ClassCastException e) {
-            System.out.println("Произошла ошибка: Вы ввели некорректный id записи.");
-        } catch (IllegalArgumentException e) {
+            System.out.println("Произошла ошибка: Неправильный формат даты рождения. Запись не обновлена.");
+        } catch (InputMismatchException e) {
+            System.out.println("Произошла ошибка: Введён некорректный ID. Запись не обновлена.");
+            scanner.nextLine();
+        } catch (IllegalArgumentException | IfListEmpty e) {
             System.out.println(e.getMessage());
         }
     }
@@ -98,11 +141,16 @@ public class PersonController {
         try {
             List<Person> personList = personService.list();
             if (personList.isEmpty()) {
-                throw new Exception("Список пользователей пуст!");
+                throw new IfListEmpty("Список пользователей пуст!");
             }
             List<Person> persons = personService.findAll();
+            System.out.print("Доступные id пользователей для удаления - введите, тот который нужен: ");
+            for (Person person : persons) {
+                System.out.print(person.getId() + " ");
+            }
+            System.out.println();
             int maxId = persons.size();
-            System.out.println("Введите id записи для удаления (от 1 до " + maxId + "):");
+            System.out.println("Всего id пользователей доступных для удаления 1 из " + maxId + "):");
             int id = scanner.nextInt();
             Person personToDelete = null;
             for (Person person : personList) {
@@ -112,22 +160,31 @@ public class PersonController {
                 }
             }
             if (personToDelete == null) {
-                throw new Exception("Ошибка: Запись с id " + id + " не найдена");
+                throw new IllegalVolumeDelete("Ошибка: Запись с id " + id + " не найдена");
             } else {
-                boolean isDeleted = personService.delete(id);
-                if (isDeleted) {
-                    personList.remove(personToDelete);
-                    System.out.println("Запись удалена");
-                    for (int i = id - 1; i < personList.size(); i++) {
-                        personList.get(i).setId(i + 1);
+                System.out.println("Вы действительно хотите удалить пользователя " + personToDelete.getName() + " под id" + personToDelete.getId()  + "?" + " (Да/Нет)");
+                String answer = scanner.next();
+                if (answer.equalsIgnoreCase("Да")) {
+                    boolean isDeleted = personService.delete(id);
+                    if (isDeleted) {
+                        personList.remove(personToDelete);
+                        System.out.println("Запись удалена");
+                        for (int i = id - 1; i < personList.size(); i++) {
+                            personList.get(i).setId(i + 1);
+                        }
+                    } else {
+                        throw new IllegalVolumeDelete("Ошибка: Не удалось удалить запись с id " + id);
                     }
+                } else if (answer.equalsIgnoreCase("Нет")) {
+                    System.out.println("Удаление отменено");
                 } else {
-                    throw new Exception("Ошибка: Не удалось удалить запись с id " + id);
+                    throw new IllegalArgumentException("Ошибка: Некорректный ответ");
                 }
             }
-        } catch (NoSuchElementException e) {
-            System.out.println("Ошибка: Введено неверное значение");
-        } catch (Exception e) {
+        } catch (InputMismatchException e) {
+            System.out.println("Ошибка: Некорректный ввод, введите число");
+            scanner.next();
+        } catch (IfListEmpty | IllegalVolumeDelete | IllegalArgumentException e) {
             System.out.println(e.getMessage());
         }
     }
@@ -135,11 +192,17 @@ public class PersonController {
         try {
             List<Person> persons = personService.list();
             if (persons.isEmpty()) {
-                throw new Exception("Список пользователей пуст!");
+                throw new IfListEmpty("Список пользователей пуст!");
             }
             System.out.println("Список всех записей:");
+            Scanner scanner = new Scanner(System.in);
+            System.out.println("Хотите отсортировать список по имени? (Да/Нет)");
+            String answer = scanner.nextLine();
+            if (answer.equals("Да")) {
+                persons.sort(Comparator.comparing(Person::getName));
+            }
             persons.forEach(System.out::println);
-        } catch (Exception e) {
+        } catch (IfListEmpty e) {
             System.out.println(e.getMessage());
         }
     }
@@ -151,8 +214,12 @@ public class PersonController {
             } else {
                 List<Person> persons = personService.findAll();
                 int maxId = persons.size();
+                System.out.print("Доступные id пользователей для поиска (от 1 до " + maxId + ")" +  ".\nID после двоеточия вы можете использовать для поиска: ");
+                for (Person person : persons) {
+                    System.out.print(person.getId() + " ");
+                }
+                System.out.println();
                 scanner.nextLine();
-                System.out.println("Введите id записи для поиска (от 1 до " + maxId + "):");
                 String id = scanner.nextLine();
                 foundPersons = personService.searchById(Integer.parseInt(id));
                 if (foundPersons.isEmpty()) {
